requirements.txt
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
mmcv==1.7.0
git+https://github.com/huggingface/diffusers
timm==0.6.12
git+https://github.com/huggingface/accelerate
tensorboard
tensorboardX
transformers==4.39.0
sentencepiece~=0.1.99
ftfy
beautifulsoup4
protobuf==3.20.2
gradio==4.1.1
yapf==0.40.1
opencv-python
bs4
einops
xformers
optimum
peft==0.6.2
pytorch-fid
trl
pillow
flash-attn
webdataset
came_pytorch
fairscale
pillow>=10.0.0
diffusers
